fix: disable btn when not click evaluate (#38)

+++ components/Promptbar/components/PromptbarSettings.tsx
@@ -17,6 +17,7 @@ import { Score } from './Score';
 
 import { ClientFactory } from '@/lib/clientFactory';
 
+
 interface Props {}
 
 export const PromptbarSettings: FC<Props> = () => {
@@ -32,6 +33,7 @@ export const PromptbarSettings: FC<Props> = () => {
   const score = state.score;
   const bestScore = state.bestScore;
   const [problem, setProblem] = useState<Problem | null>(null);
+  const [isEvaluated, setIsEvaluated] = useState<boolean>(false);
   const [evaluateLoading, setEvaluateLoading] = useState(false);
   const [submitLoading, setSubmitLoading] = useState(false);
   const { reward, isAnimating } = useReward('rewardId', 'confetti', {
@@ -48,6 +50,7 @@ export const PromptbarSettings: FC<Props> = () => {
 
   const handleEvaluate = async () => {
     setEvaluateLoading(true);
+    setIsEvaluated(true);
     const selectedConversation = state.selectedConversation;
     if (!selectedConversation) {
       alert('会話を選択してください');
@@ -75,34 +78,39 @@ export const PromptbarSettings: FC<Props> = () => {
   const router = useRouter();
 
   const handleSubmit = async () => {
-    setSubmitLoading(true);
-    const selectedConversation = state.selectedConversation;
-    if (!selectedConversation) {
-      alert('会話を選択してください');
-      setSubmitLoading(false);
-      return;
-    } else {
-      // todo: user_idを取得する
-      const submission = await prompthonClient.createSubmission({
-        user_id: 1,
-        competition_id: competitionId,
-        problem_id: problemId,
-        content: selectedConversation,
-      });
-      localStorage.setItem('tmp.submission', JSON.stringify(submission));
-      console.log(submission.score);
-      console.log(submission.content);
-      handleClearConversations();
-      setSubmitLoading(false);
-      router.push({
-        pathname: '/result',
-        query: {
-          userId: 1,
-          problemId,
-          competitionId,
-          submissionId: submission.id,
-        },
-      });
+    if (isEvaluated) {
+      const isConfirmed = window.confirm(`${bestScore}点のスコアで提出しますか？`);
+      if (isConfirmed) {
+        setSubmitLoading(true);
+        const selectedConversation = state.selectedConversation;
+        if (!selectedConversation) {
+          alert('会話を選択してください');
+          setSubmitLoading(false);
+          return;
+        } else {
+          // todo: user_idを取得する
+          const submission = await prompthonClient.createSubmission({
+            user_id: 1,
+            competition_id: competitionId,
+            problem_id: problemId,
+            content: selectedConversation,
+          });
+          localStorage.setItem('tmp.submission', JSON.stringify(submission));
+          console.log(submission.score);
+          console.log(submission.content);
+          handleClearConversations();
+          setSubmitLoading(false);
+          router.push({
+            pathname: '/result',
+            query: {
+              userId: 1,
+              problemId,
+              competitionId,
+              submissionId: submission.id,
+            },
+          });
+        }
+      }
     }
   };
 
@@ -130,8 +138,9 @@ export const PromptbarSettings: FC<Props> = () => {
         )}
       </button>
 
+        {/* className="text-sidebar flex w-[260px] h-[64px] p-3 flex-shrink-0 cursor-pointer select-none items-center gap-3 transition-colors duration-200 hover:bg-gray-500 text-white" */}
       <button
-        className="text-sidebar flex w-[260px] h-[64px] p-3 flex-shrink-0 cursor-pointer select-none items-center gap-3 transition-colors duration-200 hover:bg-gray-500 text-white"
+        className={`text-sidebar flex w-[260px] h-[64px] p-3 flex-shrink-0 cursor-pointer select-none items-center gap-3 transition-colors duration-200 ${isEvaluated ? 'hover:bg-gray-500 text-white' : 'bg-gray text-gray-600 pointer-events-none'}`}
         onClick={handleSubmit}
       >
         {submitLoading ? (
